+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+* Fixed defaced values for get network operator/sim info
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.11.9 TEST**